+++ react/utils/profile.js
@@ -1,0 +1,3 @@
+import { path } from 'ramda'
+
+export const getProfile = data => path(['getSession', 'profile'], data)