+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Redirecting to the previous route in the app when the user tries to go to Login when already logged
 
 ## [1.5.5] - 2018-09-26
 ### Fixed